Collision, isValidMovementPath, and isLegal are completed.
Collision() is designed to identify any pieces in the path of the moving piece (excluding the knight). Any collisions will disqualify the user-provided movement path as legal.

In addition, the public and private members of the Rules class were made constant where applicable. And the isValidPawnMove() was updated to check for the existence of an opposing piece on every attempted diagonal move.